Changes after discussion with Alex re status and seperating class
